+++ levels/raceEditor/splinesEditing.cpp
@@ -290,7 +290,7 @@ namespace Levels
 
 			std::vector<glm::vec2> intermediatePositions;
 
-			const int numOfSplineVertices = splineDef.complexity * (controlPoints.size() - 1 + splineDef.loop) + 1;
+			const size_t numOfSplineVertices = splineDef.complexity * (controlPoints.size() - 1 + splineDef.loop) + 1;
 
 			Tools::CubicHermiteSpline spline = splineDef.loop
 				? Tools::CubicHermiteSpline(std::move(controlPoints), Tools::CubicHermiteSpline<>::loop)
@@ -298,7 +298,7 @@ namespace Levels
 
 			intermediatePositions.reserve(numOfSplineVertices);
 
-			for (int i = 0; i < numOfSplineVertices; ++i)
+			for (size_t i = 0; i < numOfSplineVertices; ++i)
 			{
 				const float t = (float)i / (numOfSplineVertices - 1);
 				intermediatePositions.push_back(glm::vec3(spline.getInterpolation(t), 0.0f));
